// generated/src/app/api-models/codereview-output-w.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CodereviewOutputW {
  loadcount: number;
  claimNumber: string;
  lineNumber: number;
  histClaimNumber: string;
  histLineNumber: number;
  histChangedFlg: string;
  codeStatusIn: string;
  codeIn: string;
  msgCodeIn: string;
  outputPointer: string;
  codeStatusOut: string;
  pmatchCodeStatus: string;
  codeOut: string;
  msgCodeOut: string;
  pmatchMsgCode: string;
  modifier1Out: string;
  crArtificialLineNo: number;
  expandPointer: number;
  svcFromDate: Date;
  svcToDate: Date;
  codereviewServiceQty: number;
  billedAmt: number;
  histClaimNoIn: string;
  histClaimDtlIn: number;
  modifier2Out: string;
  mod1In: string;
  mod2In: string;
}